Corrected Indentation using formatter. (#1541) (#1544)
