+++ vertx/src/main/java/com/baeldung/SimpleServerVerticle.java
@@ -8,10 +8,8 @@ public class SimpleServerVerticle extends AbstractVerticle {
     @Override
     public void start(Future<Void> future) {
         vertx.createHttpServer()
-                .requestHandler(request -> {
-                    request.response()
-                            .end("Welcome to Vert.x Intro");
-                })
+                .requestHandler(
+                  r -> r.response().end("Welcome to Vert.x Intro"))
                 .listen(config().getInteger("http.port", 8080), result -> {
                     if (result.succeeded()) {
                         future.complete();
