+++ vertx/src/test/java/com/baeldung/RestServiceVerticleIntegrationTest.java
@@ -5,23 +5,23 @@ import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
-import io.vertx.core.DeploymentOptions;
+import com.baledung.rest.RestServiceVerticle;
+
 import io.vertx.core.Vertx;
-import io.vertx.core.json.JsonObject;
 import io.vertx.ext.unit.Async;
 import io.vertx.ext.unit.TestContext;
 import io.vertx.ext.unit.junit.VertxUnitRunner;
 
 @RunWith(VertxUnitRunner.class)
-public class SimpleServerVerticleTest {
+public class RestServiceVerticleIntegrationTest {
+
     private Vertx vertx;
 
     @Before
     public void setup(TestContext testContext) {
         vertx = Vertx.vertx();
 
-        vertx.deployVerticle(SimpleServerVerticle.class.getName(),
-                testContext.asyncAssertSuccess());
+        vertx.deployVerticle(RestServiceVerticle.class.getName(), testContext.asyncAssertSuccess());
     }
 
     @After
@@ -30,18 +30,17 @@ public class SimpleServerVerticleTest {
     }
 
     @Test
-    public void whenReceivedResponse_thenSuccess(TestContext testContext) {
+    public void givenId_whenReceivedArticle_thenSuccess(TestContext testContext) {
         final Async async = testContext.async();
 
         vertx.createHttpClient()
-                .getNow(8080, "localhost", "/", response -> {
+                .getNow(8080, "localhost", "/api/baeldung/articles/article/12345", response -> {
                     response.handler(responseBody -> {
                         testContext.assertTrue(responseBody.toString()
-                                .contains("Welcome"));
+                                .contains("\"id\" : \"12345\""));
                         async.complete();
                     });
                 });
     }
 
 }
-
